+++ src/core/QCheck.mli
@@ -165,7 +165,7 @@ module Gen : sig
 
   val (<$>) : ('a -> 'b) -> 'a t -> 'b t
   (** An infix synonym for {!map}
-      @since NEXT_RELEASE *)
+      @since 0.13 *)
 
   val oneof : 'a t list -> 'a t
   (** Constructs a generator that selects among a given list of generators. *)
@@ -241,15 +241,20 @@ module Gen : sig
 
   val nat : int t (** Generates small natural numbers. *)
 
-  val big_nat : int t (** Generates natural numbers, possibly large. @since 0.10 *)
+  val big_nat : int t
+  (** Generates natural numbers, possibly large.
+      @since 0.10 *)
 
-  val neg_int : int t (** Generates non-strictly negative integers (0 included). *)
+  val neg_int : int t
+  (** Generates non-strictly negative integers (0 included). *)
 
   val pint : int t (** Generates non-strictly positive integers uniformly (0 included). *)
 
   val int : int t (** Generates integers uniformly. *)
 
-  val small_nat : int t (** Small integers (< 100) @since 0.5.1 *)
+  val small_nat : int t
+  (** Small integers (< 100)
+      @since 0.5.1 *)
 
   val small_int : int t
   (** Small UNSIGNED integers, for retrocompatibility.
@@ -266,7 +271,7 @@ module Gen : sig
 
   val int_range : int -> int -> int t
   (** Uniform integer generator producing integers within [low,high].
-      @raise Invalid_argument if [low > high] or if the range is larger than [max_int]. *)
+      @raise Invalid_argument if [low > high]. *)
 
   val graft_corners : 'a t -> 'a list -> unit -> 'a t
   (** [graft_corners gen l ()] makes a new generator that enumerates
@@ -311,9 +316,12 @@ module Gen : sig
 
   val triple : 'a t -> 'b t -> 'c t -> ('a * 'b * 'c) t (** Generates triples. *)
 
-  val quad : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t (** Generates quadruples. @since 0.5.1 *)
+  val quad : 'a t -> 'b t -> 'c t -> 'd t -> ('a * 'b * 'c * 'd) t
+  (** Generates quadruples.
+      @since 0.5.1 *)
 
-  val char : char t (** Generates characters upto character code 255. *)
+  val char : char t
+  (** Generates characters upto character code 255. *)
 
   val printable : char t (** Generates printable characters. *)
 
@@ -322,7 +330,7 @@ module Gen : sig
   val char_range : char -> char -> char t
   (** Generates chars between the two bounds, inclusive.
       Example: [char_range 'a' 'z'] for all lower case ascii letters.
-      @since NEXT_RELEASE *)
+      @since 0.13 *)
 
   val string_size : ?gen:char t -> int t -> string t
   (** Builds a string generator from a (non-negative) size generator.
@@ -349,6 +357,22 @@ module Gen : sig
   val small_list : 'a t -> 'a list t
   (** Generates lists of small size (see {!small_nat}).
       @since 0.5.3 *)
+
+  val flatten_l : 'a t list -> 'a list t
+  (** Generate a list of elements from individual generators
+      @since 0.13 *)
+
+  val flatten_a : 'a t array -> 'a array t
+  (** Generate an array of elements from individual generators
+      @since 0.13 *)
+
+  val flatten_opt : 'a t option -> 'a option t
+  (** Generate an option from an optional generator
+      @since 0.13 *)
+
+  val flatten_res : ('a t, 'e) result -> ('a,'e) result t
+  (** Generate a result from [Ok g], an error from [Error e]
+      @since 0.13 *)
 
   val small_array : 'a t -> 'a array t
   (** Generates arrays of small size (see {!small_nat}).
@@ -403,6 +427,9 @@ module Gen : sig
 
   val generate1 : ?rand:Random.State.t -> 'a t -> 'a
   (** [generate1 g] generates one instance of [g]. *)
+
+  include Qcheck_ops.S with type 'a t_let := 'a t
+  (** @since 0.15 *)
 end
 
 (** {2 Pretty printing} *)
@@ -479,6 +506,9 @@ module Iter : sig
 
   val flatten : 'a t t -> 'a t
   (** @since 0.8 *)
+
+  include Qcheck_ops.S with type 'a t_let := 'a t
+  (** @since 0.15 *)
 end
 
 (** {2 Shrink Values}
@@ -500,6 +530,12 @@ module Shrink : sig
   val char : char t (** @since 0.6 *)
 
   val int : int t
+
+  val int32 : int32 t
+  (** @since 0.14 *)
+
+  val int64 : int64 t
+  (** @since 0.14 *)
 
   val option : 'a t -> 'a option t
 
@@ -618,7 +654,7 @@ type 'a arbitrary = private {
     {b NOTE} the collect field is unstable and might be removed, or
     moved into {!Test}.
 
-    Made private @since 0.8
+    Made private since 0.8
 *)
 
 val make :
@@ -674,8 +710,12 @@ val gen : 'a arbitrary -> 'a Gen.t
 module TestResult : sig
   type 'a counter_ex = {
     instance: 'a; (** The counter-example(s) *)
+
     shrink_steps: int; (** How many shrinking steps for this counterex *)
-    msg_l: string list; (** messages. @since 0.7 *)
+
+    msg_l: string list;
+    (** messages.
+        @since 0.7 *)
   }
 
   type 'a failed_state = 'a counter_ex list
@@ -793,7 +833,7 @@ module Test : sig
       See {!make_cell} for a description of the parameters.
   *)
 
-  (** {6 Running the test} *)
+  (** {3 Running the test} *)
 
   exception Test_fail of string * string list
   (** Exception raised when a test failed, with the list of counter-examples.
@@ -925,7 +965,7 @@ val find_example_gen :
     @raise No_example_found if no example was found within [count] tries.
     @since 0.6 *)
 
-(** {2 Combinators for {!arbitrary}} *)
+(** {2 Combinators for arbitrary} *)
 
 val choose : 'a arbitrary list -> 'a arbitrary
 (** Choose among the given list of generators. The list must not
@@ -1087,15 +1127,19 @@ val fun1_unsafe : 'a arbitrary -> 'b arbitrary -> ('a -> 'b) arbitrary
     - when given the same argument (as decided by Pervasives.(=)), it returns the same value
     - it never does side effects, like printing or never raise exceptions etc.
     The functions generated are really printable.
-    renamed from {!fun1}. @since 0.6
-    @deprecated use {!fun_} instead. @since 0.6
+
+    renamed from {!fun1} since 0.6
+
+    @deprecated use {!fun_} instead.
+
+    @since 0.6
 *)
 
 val fun2_unsafe : 'a arbitrary -> 'b arbitrary -> 'c arbitrary -> ('a -> 'b -> 'c) arbitrary
 (** Generator of functions of arity 2. The remark about [fun1] also apply
     here.
-    renamed from {!fun2}. @since 0.6
-    @deprecated use {!fun_} instead. @since 0.6
+    renamed from {!fun2} since 0.6
+    @deprecated use {!fun_} instead since 0.6
 *)
 
 type _ fun_repr