+++ src/runner/QCheck_base_runner.ml
@@ -4,10 +4,6 @@ copyright (c) 2013-2017, Guillaume Bury, Simon Cruanes, Vincent Hugot, Jan Midtg
 all rights reserved.
 *)
 
-let ps = print_string
-let va = Printf.sprintf
-let pf = Printf.printf
-
 module Color = struct
   let fpf = Printf.fprintf
   type color =
@@ -50,15 +46,13 @@ module Color = struct
     else output_string out s
 end
 
-let separator1 = "\027[K" ^ (String.make 79 '\\')
-let separator2 = String.make 79 '/'
-
 let seed = ref ~-1
 let st = ref None
 
-let set_seed_ s =
+let set_seed_ ~colors s =
   seed := s;
-  Printf.printf "%srandom seed: %d\n%!" Color.reset_line s;
+  if colors then Printf.printf "%srandom seed: %d\n%!" Color.reset_line s
+  else Printf.printf "random seed: %d\n%!" s;
   let state = Random.State.make [| s |] in
   st := Some state;
   state
@@ -72,19 +66,21 @@ let get_time_between_msg () = !time_between_msg
 
 let set_time_between_msg f = time_between_msg := f
 
-let set_seed s = ignore (set_seed_ s)
+let set_seed s = ignore (set_seed_ ~colors:false s)
 
-let setup_random_state_ () =
+let setup_random_state_ ~colors () =
   let s = if !seed = ~-1 then (
       Random.self_init ();  (* make new, truly random seed *)
       Random.int (1 lsl 29);
     ) else !seed in
-  set_seed_ s
+  set_seed_ ~colors s
 
 (* initialize random generator from seed (if any) *)
-let random_state () = match !st with
+let random_state_ ~colors () = match !st with
   | Some st -> st
-  | None -> setup_random_state_ ()
+  | None -> setup_random_state_ ~colors ()
+
+let random_state() = random_state_ ~colors:false ()
 
 let verbose, set_verbose =
   let r = ref false in
@@ -93,6 +89,14 @@ let verbose, set_verbose =
 let long_tests, set_long_tests =
   let r = ref false in
   (fun () -> !r), (fun b -> r := b)
+
+let debug_shrink, set_debug_shrink =
+  let r = ref None in
+  (fun () -> !r), (fun s -> r := Some (open_out s))
+
+let debug_shrink_list, set_debug_shrink_list =
+  let r = ref [] in
+  (fun () -> !r), (fun b -> r := b :: !r)
 
 module Raw = struct
   type ('b,'c) printer = {
@@ -108,18 +112,21 @@ module Raw = struct
     cli_rand : Random.State.t;
     cli_slow_test : int; (* how many slow tests to display? *)
     cli_colors: bool;
+    cli_debug_shrink : out_channel option;
+    cli_debug_shrink_list : string list;
   }
 
   (* main callback for individual tests
      @param verbose if true, print statistics and details
      @param print_res if true, print the result on [out] *)
-  let callback ~verbose ~print_res ~print name cell result =
+  let callback ~colors ~verbose ~print_res ~print name cell result =
     let module R = QCheck.TestResult in
     let module T = QCheck.Test in
     let arb = T.get_arbitrary cell in
+    let reset_line = if colors then Color.reset_line else "\n" in
     if verbose then (
       print.info "%slaw %s: %d relevant cases (%d total)\n"
-        Color.reset_line name result.R.count result.R.count_gen;
+        reset_line name result.R.count result.R.count_gen;
       begin match QCheck.TestResult.collect result with
         | None -> ()
         | Some tbl ->
@@ -131,11 +138,11 @@ module Raw = struct
       match result.R.state with
         | R.Success -> ()
         | R.Failed {instances=l} ->
-          print.fail "%s%s\n" Color.reset_line (T.print_fail arb name l);
+          print.fail "%s%s\n" reset_line (T.print_fail arb name l);
         | R.Failed_other {msg} ->
-          print.fail "%s%s\n" Color.reset_line (T.print_fail_other name ~msg);
+          print.fail "%s%s\n" reset_line (T.print_fail_other name ~msg);
         | R.Error {instance; exn; backtrace} ->
-          print.err "%s%s\n" Color.reset_line
+          print.err "%s%s\n" reset_line
             (T.print_error ~st:backtrace arb name (instance,exn));
     )
 
@@ -165,13 +172,16 @@ module Raw = struct
           ; "--seed", Arg.Set_int seed, " set random seed (to repeat tests)"
           ; "--long", Arg.Unit set_long_tests, " run long tests"
           ; "-bt", Arg.Unit set_backtraces, " enable backtraces"
+          ; "--debug-shrink", Arg.String set_debug_shrink, " enable shrinking debug to <file>"
+          ; "--debug-shrink-list", Arg.String set_debug_shrink_list, " filter test to debug shrinking on"
           ]
       ) in
     Arg.parse_argv argv options (fun _ ->()) "run qtest suite";
-    let cli_rand = setup_random_state_ () in
+    let cli_rand = setup_random_state_ ~colors:!colors () in
     { cli_verbose=verbose(); cli_long_tests=long_tests(); cli_rand;
       cli_print_list= !print_list; cli_slow_test= !slow;
-      cli_colors= !colors; }
+      cli_colors= !colors; cli_debug_shrink = debug_shrink();
+      cli_debug_shrink_list = debug_shrink_list(); }
 end
 
 open Raw
@@ -189,32 +199,88 @@ type counter = {
 type res =
   | Res : 'a QCheck.Test.cell * 'a QCheck.TestResult.t -> res
 
+type handler = {
+  handler : 'a. 'a QCheck.Test.handler;
+}
+
+type handler_gen =
+  colors:bool ->
+  debug_shrink:(out_channel option) ->
+  debug_shrink_list:(string list) ->
+  size:int -> out:out_channel -> verbose:bool -> counter -> handler
+
 let pp_counter ~size out c =
   let t = Unix.gettimeofday () -. c.start in
   Printf.fprintf out "%*d %*d %*d %*d / %*d %7.1fs"
     size c.gen size c.errored size c.failed
     size c.passed size c.expected t
 
-let handler ~size ~out ~verbose c name _ r =
-  let st = function
-    | QCheck.Test.Generating    -> "generating"
-    | QCheck.Test.Collecting _  -> "collecting"
-    | QCheck.Test.Testing _     -> "   testing"
-    | QCheck.Test.Shrunk (i, _) ->
-      Printf.sprintf "shrinking: %4d" i
-    | QCheck.Test.Shrinking (i, j, _) ->
-      Printf.sprintf "shrinking: %4d.%04d" i j
+let debug_shrinking_counter_example cell out x =
+  match (QCheck.Test.get_arbitrary cell).QCheck.print with
+  | None -> Printf.fprintf out "<no printer provided>"
+  | Some print -> Printf.fprintf out "%s" (print x)
+
+let debug_shrinking_size cell out x =
+  match (QCheck.Test.get_arbitrary cell).QCheck.small with
+  | None -> ()
+  | Some f -> Printf.fprintf out ", size %d" (f x)
+
+let debug_shrinking_choices_aux ~colors out name i cell x =
+  Printf.fprintf out "\n~~~ %a %s\n\n"
+    (Color.pp_str_c ~colors `Cyan) "Shrink" (String.make 69 '~');
+  Printf.fprintf out
+    "Test %s sucessfully shrunk counter example (step %d%a) to:\n\n%a\n%!"
+    name i
+    (debug_shrinking_size cell) x
+    (debug_shrinking_counter_example cell) x
+
+let debug_shrinking_choices
+    ~colors ~debug_shrink ~debug_shrink_list name cell i x =
+  match debug_shrink with
+  | None -> ()
+  | Some out ->
+    begin match debug_shrink_list with
+      | [] ->
+        debug_shrinking_choices_aux ~colors out name i cell x
+      | l when List.mem name l ->
+        debug_shrinking_choices_aux ~colors out name i cell x
+      | _ -> ()
+    end
+
+
+let default_handler
+  ~colors ~debug_shrink ~debug_shrink_list
+  ~size ~out ~verbose c =
+  let handler name cell r =
+    let st = function
+      | QCheck.Test.Generating    -> "generating"
+      | QCheck.Test.Collecting _  -> "collecting"
+      | QCheck.Test.Testing _     -> "   testing"
+      | QCheck.Test.Shrunk (i, _) ->
+        Printf.sprintf "shrinking: %4d" i
+      | QCheck.Test.Shrinking (i, j, _) ->
+        Printf.sprintf "shrinking: %4d.%04d" i j
+    in
+    (* debug shrinking choices *)
+    begin match r with
+      | QCheck.Test.Shrunk (i, x) ->
+          debug_shrinking_choices
+          ~colors ~debug_shrink ~debug_shrink_list name cell i x
+      | _ ->
+        ()
+    end;
+    (* use timestamps for rate-limiting *)
+    let now=Unix.gettimeofday() in
+    if verbose && now -. !last_msg > get_time_between_msg () then (
+      last_msg := now;
+      Printf.fprintf out "%s[ ] %a %s (%s)%!"
+        (if colors then Color.reset_line else "\n")
+        (pp_counter ~size) c name (st r)
+    )
   in
-  (* use timestamps for rate-limiting *)
-  let now=Unix.gettimeofday() in
-  if verbose && now -. !last_msg > get_time_between_msg () then (
-    last_msg := now;
-    Printf.fprintf out "%s[ ] %a %s (%s)%!"
-      Color.reset_line (pp_counter ~size) c name (st r)
-  )
+  { handler; }
 
-
-let step ~size ~out ~verbose c name _ _ r =
+let step ~colors ~size ~out ~verbose c name _ _ r =
   let aux = function
     | QCheck.Test.Success -> c.passed <- c.passed + 1
     | QCheck.Test.Failure -> c.failed <- c.failed + 1
@@ -227,15 +293,15 @@ let step ~size ~out ~verbose c name _ _ r =
   if verbose && now -. !last_msg > get_time_between_msg () then (
     last_msg := now;
     Printf.fprintf out "%s[ ] %a %s%!"
-      Color.reset_line (pp_counter ~size) c name
+      (if colors then Color.reset_line else "\n") (pp_counter ~size) c name
   )
 
-let callback ~size ~out ~verbose ~colors c name _ _ =
-  let pass = c.failed = 0 && c.errored = 0 in
+let callback ~size ~out ~verbose ~colors c name _ r =
+  let pass = QCheck.TestResult.is_success r in
   let color = if pass then `Green else `Red in
   if verbose then (
     Printf.fprintf out "%s[%a] %a %s\n%!"
-      Color.reset_line
+      (if colors then Color.reset_line else "\n")
       (Color.pp_str_c ~bold:true ~colors color) (if pass then "✓" else "✗")
       (pp_counter ~size) c name
   )
@@ -278,12 +344,14 @@ let print_success ~colors out cell r =
         (Color.pp_str_c ~colors `Yellow) "Warning" (String.make 68 '!')
         (QCheck.Test.get_name cell) msg)
     (QCheck.TestResult.warnings r);
+
+  if QCheck.TestResult.stats r <> []  then
+     Printf.fprintf out
+       "\n+++ %a %s\n%!"
+       (Color.pp_str_c ~colors `Blue) ("Stats for " ^ QCheck.Test.get_name cell)
+       (String.make 56 '+');
   List.iter
-    (fun st ->
-       Printf.fprintf out
-         "\n+++ %a %s\n\nStat for test %s:\n\n%s%!"
-        (Color.pp_str_c ~colors `Blue) "Stat" (String.make 68 '+')
-        (QCheck.Test.get_name cell) (QCheck.Test.print_stat st))
+    (fun st -> Printf.fprintf out "\n%s%!" (QCheck.Test.print_stat st))
     (QCheck.TestResult.stats r);
   ()
 
@@ -309,7 +377,11 @@ let print_error ~colors out cell c_ex exn bt =
   print_messages ~colors out cell c_ex.QCheck.TestResult.msg_l
 
 let run_tests
-    ?(colors=true) ?(verbose=verbose()) ?(long=long_tests()) ?(out=stdout) ?(rand=random_state()) l =
+    ?(handler=default_handler)
+    ?(colors=true) ?(verbose=verbose()) ?(long=long_tests())
+    ?(debug_shrink=debug_shrink()) ?(debug_shrink_list=debug_shrink_list())
+    ?(out=stdout) ?rand l =
+  let rand = match rand with Some x -> x | None -> random_state_ ~colors () in
   let module T = QCheck.Test in
   let module R = QCheck.TestResult in
   let pp_color = Color.pp_str_c ~bold:true ~colors in
@@ -329,10 +401,12 @@ let run_tests
     } in
     if verbose then
       Printf.fprintf out "%s[ ] %a %s%!"
-        Color.reset_line (pp_counter ~size) c (T.get_name cell);
+        (if colors then Color.reset_line else "")
+        (pp_counter ~size) c (T.get_name cell);
     let r = QCheck.Test.check_cell ~long ~rand
-        ~handler:(handler ~size ~out ~verbose c)
-        ~step:(step ~size ~out ~verbose c)
+        ~handler:(handler ~colors ~debug_shrink ~debug_shrink_list
+                    ~size ~out ~verbose c).handler
+        ~step:(step ~colors ~size ~out ~verbose c)
         ~call:(callback ~size ~out ~verbose ~colors c)
         cell
     in